Fix error messages and codes for DELETE /data/aces
[#139762095] User should be able to create/delete ACEs via /aces endpoint based on 'write_acl' operation ACE

+++ src/main/java/io/pivotal/security/repository/AccessEntryRepository.java
@@ -15,5 +15,5 @@ public interface AccessEntryRepository extends JpaRepository<AccessEntryData, UU
   AccessEntryData findByCredentialNameAndActor(CredentialName credentialName, String actor);
 
   @Transactional
-  int deleteByCredentialNameAndActor(CredentialName credentialName, String actor);
+  long deleteByCredentialNameAndActor(CredentialName credentialName, String actor);
 }
